refactored http utility functions into their own package
