+++ pkg/client/http/errors.go
@@ -1,4 +1,4 @@
-package client
+package http
 
 import (
 	"errors"
